refactor: move AboutContactCard into GenericPagesComponents
