+++ src/pages/about-us/board.tsx
@@ -6,7 +6,7 @@ import { AboutBoardPage } from "@/common-lib/cms-types";
 import Layout from "@/components/SharedComponents/Layout";
 import MaxWidth from "@/components/SharedComponents/MaxWidth";
 import Card from "@/components/SharedComponents/Card";
-import AboutContactCard from "@/components/AboutContactCard";
+import AboutContactCard from "@/components/GenericPagesComponents/AboutContactCard";
 import Typography, {
   Heading,
   Hr,
