+++ src/pages/about-us/work-with-us.tsx
@@ -6,7 +6,7 @@ import { AboutWorkWithUsPage } from "@/common-lib/cms-types";
 import Layout from "@/components/SharedComponents/Layout";
 import MaxWidth from "@/components/SharedComponents/MaxWidth";
 import Card from "@/components/SharedComponents/Card";
-import AboutContactCard from "@/components/AboutContactCard";
+import AboutContactCard from "@/components/GenericPagesComponents/AboutContactCard";
 import Typography, { Heading } from "@/components/SharedComponents/Typography";
 import ButtonAsLink from "@/components/SharedComponents/Button/ButtonAsLink";
 import Flex from "@/components/SharedComponents/Flex";
